Support multiapp pm in deploy APIs
sl-deploy was resolving a service name to the deploy URL, but this is
commonly needed code: the deploy APIs can't be used if you don't know
the URl to deploy to, and its the mesh client that can tell us. Refactor
so the APIs do the resolution.

In the process, fix local deploy, which didn't get multiapp support when
it was added for packfile and git deployments.